Tue, May 6, 2025, 7:48 AM -04:00

+++ interpreter/interpreter.py
@@ -4,11 +4,11 @@ x,y,z = operation.split(" ")
 
 match y:
     case "+":
-        print(x+z)
+        print(float(x+z))
     case "-":
-        print(x-z)
+        print(float(x-z))
     case"*":
-        print(x*z)
+        print(float(x*z))
     case"/":
-        print(x/z)
+        print(float(x/z))
